skillvector-hr/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary>=2.9.9
pgvector==0.2.4
sentence-transformers>=2.3.1
pdfplumber==0.7.6
python-docx>=1.1.0
spacy>=3.7.2
rapidfuzz>=3.0.0
numpy>=1.26.0
scikit-learn>=1.3.2
openai>=1.7.0
google-generativeai>=0.3.2
transformers>=4.37.0
torch==2.2.0
Flask-Executor==1.0.0
Flask-WTF==1.2.1
groq>=0.4.2
